requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
cloudscraper==1.2.60
garminconnect==0.1.45
idna==3.3
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.1
requests-toolbelt==0.9.1
urllib3==1.26.11
